~updates~
t t t t, why you skip valid phys check?

it's still rendering the damage cracks where the "removed" entities used to be. for now i'm just moving the props into the corner of the map.

ooh can we get the render to show up on the clientside props? thatd be great.